Fix SCC resource to only affect KIC pods
Priority of 20 overrides the max priority of the default: AnyPid. This
leads to the SCC binding to all pods created in the same namespace
rather than just KIC pods.

Priority is now changed to the default value so if a new pod is created,
it takes the default SCC rather than the SCC of the operator.