Fix SCC resource to only affect KIC pods
* Remove SCC group. Fix SCC user reference
* Improve test to print diff upon mismatch

Priority of 20 overrides the max priority of the default: restricted. This
leads to the SCC binding to all pods created in the same namespace as KIC pods
rather than just KIC pods.

Priority is now changed to the default value so if a new pod is created,
it takes the default SCC rather than the SCC of the operator.