Fix leader election for IC pods
- The operator now configures the leader-election-lock-name cli
argument of the IC with the value based on the name of the
NginxIngressController resource. This allows delpoying multiple
ICs in the same namespace.
- The documentation and type of the spec.EnableLeaderElection is
updated: it is now a pointer with the default value 'true'. This
matches the default value of the corresponding IC cli argument.
The previous behavior was also incorrect - the operator would always
enable the leader election, no matter the value of
spec.EnableLeaderElection.